adding spheres

+++ src/gpu_stuff/gpu_parameters.h
@@ -10,7 +10,7 @@ struct GpuParameters {
     float beta = 0.25f;
     float overlap_factor = 4.0f;
 
-    int max_spheres = 1000000000000000;//small for now (not small anymore)
+    long int max_spheres = 1000000000000000;//small for now (not small anymore)
     int max_growth_fronts = 10000;
     int candidates_per_front = 1024;
     int max_steps = 100000000;
